CMS-3399 Create tests for ContentBrowsePanel_GridPanel_Spec

+++ src/main/java/com/enonic/autotests/pages/Application.java
@@ -17,7 +17,7 @@ public class Application
 
     public static final long IMPLICITLY_WAIT = 4l;
 
-    public static final long DEFAULT_IMPLICITLY_WAIT = 2l;
+    public static final int DEFAULT_IMPLICITLY_WAIT = 2;
 
     public static final String APP_SPACE_ADMIN_FRAME_XPATH = "//iframe[contains(@src,'space-manager')]";
